TP 458445 - [USD I/O] Fixed narrowing conversion warning/error
Summary:
We are having a narrowing conversion warning (Linux) / error (Windows) that
can be fixed using a `static_cast`.

+++ lib/usdKatana/utils.cpp
@@ -1658,7 +1658,9 @@ FnKat::Attribute PxrUsdKatanaUtils::ApplySkinningToPoints(
                     std::vector<float> attrVec(skinnedPoints.size() * 3);
                     PxrUsdKatanaUtils::ConvertArrayToVector(skinnedPoints,
                                                             &attrVec);
-                    skinnedPointsAttr = {attrVec.data(), attrVec.size(), 3};
+                    skinnedPointsAttr = {attrVec.data(),
+                                         static_cast<int64_t>(attrVec.size()),
+                                         3};
                 }
             }
         }
